changed "checkout _type" to PositiveSmallIntegerField
git-svn-id: http://django-lfs.googlecode.com/svn/trunk/lfs@161 70813176-cce2-11dd-87b6-9d1a99559f42

+++ core/models.py
@@ -99,8 +99,8 @@ class Shop(models.Model):
     
     countries = models.ManyToManyField(Country, verbose_name=_(u"Countries"), related_name="shops")
     default_country = models.ForeignKey(Country, verbose_name=_(u"Default country"))
-    
-    checkout_type = models.IntegerField(_(u"Checkout type"), choices=CHECKOUT_TYPES, default=CHECKOUT_TYPE_SELECT)
+        
+    checkout_type = models.PositiveSmallIntegerField(_(u"Checkout type"), choices=CHECKOUT_TYPES, default=CHECKOUT_TYPE_SELECT)
     
     class Meta:
         permissions = (("manage_shop", "Manage shop"),)
